Test.Integration.Faucet

Documentation

data Faucet

An opaque Faucet type from which one can get a wallet with funds

Constructors

Faucet

Fields

shelley :: MVar [Mnemonic 15]

icarus :: MVar [Mnemonic 15]

random :: MVar [Mnemonic 12]

reward :: MVar [Mnemonic 24]

ma :: MVar [Mnemonic 24]

txBuilder :: MVar [(Address, Coin) -> IO ByteString]

class NextWallet (scheme :: Symbol)

Get the next faucet wallet. Requires the initFaucet to be called in order to get a hand on a Faucet.

Minimal complete definition

nextWallet

nextWallet :: NextWallet scheme => Faucet -> IO (Mnemonic (MnemonicSize scheme))

nextTxBuilder :: Faucet -> IO ((Address, Coin) -> IO ByteString)

Get a raw transaction builder. It constructs and sign a transaction via an private key that is owned "externally". Returns a bytes string ready to be sent to a node.

Faucets

seqMnemonics :: [Mnemonic 15]

icaMnemonics :: [Mnemonic 15]

rndMnemonics :: [Mnemonic 12]

mirMnemonics :: [Mnemonic 24]

maMnemonics :: [Mnemonic 24]

Dust wallets

bigDustWallet :: Mnemonic 15

A special Shelley Wallet with 200 UTxOs where 100 of them are 1 ADA

onlyDustWallet :: Mnemonic 15

A special wallet with only dust

Sea horses

seaHorseTokenName :: Int -> TokenName

seaHorsePolicyId :: TokenPolicyId

Integration test funds

shelleyIntegrationTestFunds :: [(Address, Coin)]

byronIntegrationTestFunds :: [(Address, Coin)]

maryIntegrationTestAssets

Arguments

:: Coin
Amount of ada in each bundle

-> [(Address, (TokenBundle, [(String, String)]))]

A list of addresses, and assets to be provisioned there.

Beside the assets, there is a list of (signing key, verification key hash), so that they can be minted by the faucet.

seaHorseTestAssets

Arguments

:: Int
Number of sea horses per Address

-> Coin
The Coin value for each TokenBundle

-> [Address]

-> [(Address, (TokenBundle, [(String, String)]))]

Create n unique SeaHorse tokens for each provided Address.

The result can be used for minting using the cli-based faucet.

hwWalletFunds :: [(Text, Coin)]

Internals

genByronFaucets :: FilePath -> [Mnemonic 12] -> IO [[Text]]

Generate faucets addresses and mnemonics to a file.

>>> genMnemonics 100 >>= genByronFaucets "byron-faucets.yaml"

genShelleyFaucets :: FilePath -> [Mnemonic 15] -> IO [[Text]]

Generate faucet addresses and mnemonics to a file.

>>> genMnemonics 100 >>= genShelleyFaucets "shelley-faucets.yaml"

genMAFaucets :: FilePath -> [Mnemonic 24] -> IO [[Text]]

Generate faucet addresses and mnemonics to a file.

>>> genMnemonics 100 >>= genMAFaucets "ma-faucets.yaml"

genMnemonics :: forall mw ent csz. (ValidMnemonicSentence mw, ValidEntropySize ent, ValidChecksumSize ent csz, ent ~ EntropySize mw, mw ~ MnemonicWords ent) => Int -> IO [Mnemonic mw]

genShelleyAddresses :: SomeMnemonic -> [Address]

genRewardAccounts :: Mnemonic 24 -> [XPub]