Test.Integration.Framework.Context

Documentation

data Context

Context for integration tests.

Constructors

Context

Fields

_cleanup :: IO ()

A cleanup action.

_manager :: (URI, Manager)

The underlying base URL and manager used by the wallet client.

_walletPort :: Port "wallet"

Server TCP port.

_faucet :: Faucet

Provides access to funded wallets.

_moveRewardsToScript :: (ByteString, Coin) -> IO ()

A function to inject rewards into some stake address.

_feeEstimator :: TxDescription -> (Natural, Natural)

A fee estimator.

_networkParameters :: NetworkParameters

Blockchain parameters for the underlying chain.

_poolGarbageCollectionEvents :: IORef [PoolGarbageCollectionEvent]

The complete list of pool garbage collection events. Most recent events are stored at the head of the list.

_mainEra :: ApiEra

The main era the tests are expected to run on. Allows tests to make era-specific assertions.

_smashUrl :: Text

Base URL of the mock smash server.

_mintSeaHorseAssets :: Int -> Int -> Coin -> [Address] -> IO ()

TODO: Remove once we can unify cardano-wallet-core-integration and cardano-wallet:integration, or when the wallet supports minting.

Cannot be used by several tests at a time. (!)

data PoolGarbageCollectionEvent

Records the parameters and return value of a single call to the removeRetiredPools operation of DBLayer.

Constructors

PoolGarbageCollectionEvent

Fields

poolGarbageCollectionEpochNo :: EpochNo

The epoch number parameter.

poolGarbageCollectionCertificates :: [PoolRetirementCertificate]

The pools that were removed from the database.

data TxDescription

Describe a transaction in terms of its inputs and outputs.

Constructors

DelegDescription DelegationAction

PaymentDescription

Fields

nInputs :: Int

nOutputs :: Int

nChanges :: Int