Test.Integration.Framework.DSL

Documentation

data Context

Context for integration tests.

Constructors

Context

Fields

_cleanup :: IO ()

A cleanup action.

_manager :: (URI, Manager)

The underlying base URL and manager used by the wallet client.

_walletPort :: Port "wallet"

Server TCP port.

_faucet :: Faucet

Provides access to funded wallets.

_moveRewardsToScript :: (ByteString, Coin) -> IO ()

A function to inject rewards into some stake address.

_feeEstimator :: TxDescription -> (Natural, Natural)

A fee estimator.

_networkParameters :: NetworkParameters

Blockchain parameters for the underlying chain.

_poolGarbageCollectionEvents :: IORef [PoolGarbageCollectionEvent]

The complete list of pool garbage collection events. Most recent events are stored at the head of the list.

_mainEra :: ApiEra

The main era the tests are expected to run on. Allows tests to make era-specific assertions.

_smashUrl :: Text

Base URL of the mock smash server.

_mintSeaHorseAssets :: Int -> Int -> Coin -> [Address] -> IO ()

TODO: Remove once we can unify cardano-wallet-core-integration and cardano-wallet:integration, or when the wallet supports minting.

Cannot be used by several tests at a time. (!)

data MnemonicLength

Constructors

M9

M12

M15

M18

M21

M24

data TxDescription

Describe a transaction in terms of its inputs and outputs.

Constructors

DelegDescription DelegationAction

PaymentDescription

Fields

nInputs :: Int

nOutputs :: Int

nChanges :: Int

Steps

request

Arguments

:: forall a m s. (FromJSON a, MonadIO m, MonadUnliftIO m, HasType (URI, Manager) s)

=> s

-> (Method, Text)
HTTP method and request path

-> Headers
Request headers

-> Payload
Request body

-> m (Status, Either RequestException a)

Makes a request to the API and decodes the response.

rawRequest

Arguments

:: forall m s. (MonadIO m, MonadUnliftIO m, HasType (URI, Manager) s)

=> s

-> (Method, Text)
HTTP method and request path

-> Headers
Request headers

-> Payload
Request body

-> m (Status, Either RequestException ByteString)

Like request, but does not attempt to deserialize the response.

unsafeRequest :: forall a m. (FromJSON a, MonadIO m, MonadUnliftIO m) => Context -> (Method, Text) -> Payload -> m (Status, a)

Makes a request to the API, but throws if it fails.

unsafeResponse :: (Status, Either RequestException a) -> a

Expectations

expectPathEventuallyExist :: HasCallStack => FilePath -> IO ()

A file is eventually created on the given location

expectSuccess :: (HasCallStack, MonadIO m) => (s, Either RequestException a) -> m ()

Expect a successful response, without any further assumptions.

expectError :: (HasCallStack, MonadIO m, Show a) => (s, Either RequestException a) -> m ()

Expect an error response, without any further assumptions.

expectErrorMessage :: (HasCallStack, MonadIO m, Show a) => String -> (s, Either RequestException a) -> m ()

Expect an error response, without any further assumptions.

expectField :: (HasCallStack, MonadIO m, Show a) => Traversal' s a -> (a -> Expectation) -> (Status, Either RequestException s) -> m ()

expectListField :: (HasCallStack, MonadIO m, Show a) => Int -> Traversal' s a -> (a -> Expectation) -> (Status, Either RequestException [s]) -> m ()

expectListSize :: (HasCallStack, MonadIO m, Foldable xs) => Int -> (Status, Either RequestException (xs a)) -> m ()

Expects data list returned by the API to be of certain length

expectListSizeSatisfy :: (HasCallStack, MonadIO m, Foldable xs) => (Int -> Bool) -> (Status, Either RequestException (xs a)) -> m ()

Expects data list returned by the API to be of certain length

expectResponseCode :: (HasCallStack, MonadUnliftIO m, Show a) => Status -> (Status, a) -> m ()

Expect a given response code on the response.

expectValidJSON :: forall m a. (HasCallStack, FromJSON a, MonadIO m) => Proxy a -> String -> m a

Expects a given string to be a valid JSON output corresponding to some given data-type a. Returns this type if successful.

expectCliField :: (HasCallStack, MonadIO m, Show a) => Lens' s a -> (a -> Expectation) -> s -> m ()

expectCliListField :: (HasCallStack, MonadIO m, Show a) => Int -> Lens' s a -> (a -> Expectation) -> [s] -> m ()

expectWalletUTxO :: (HasCallStack, MonadIO m) => [Natural] -> Either RequestException ApiUtxoStatistics -> m ()

Expects wallet UTxO statistics from the request to be equal to pre-calculated statistics.

between :: (Ord a, Show a) => (a, a) -> a -> Expectation

counterexample :: (MonadIO m, MonadUnliftIO m, HasCallStack) => String -> m a -> m a

Can be used to add context to a HUnitFailure.

>>> counterexample (show response) (0 `shouldBe` 3)
>>> (Status {statusCode = 200, statusMessage = "OK"},Right [])
>>> expected: 3
>>> but got: 0

(.>=) :: (Ord a, Show a) => a -> a -> Expectation

(.<=) :: (Ord a, Show a) => a -> a -> Expectation

(.>) :: (Ord a, Show a) => a -> a -> Expectation

(.<) :: (Ord a, Show a) => a -> a -> Expectation

verify :: (Show a, MonadUnliftIO m) => a -> [a -> m ()] -> m ()

Applies the value a to all assertions in the given sequence.

If any of the assertions fail, a is shown as the counter-example text.

verifyMsg :: (Show a, MonadUnliftIO m) => String -> a -> [a -> m ()] -> m ()

Applies the value a to all assertions in the given sequence.

Like verify, but the counterexample shows a description of what conditions were being checked.

data Headers

The headers of the request

Constructors

Headers RequestHeaders

Default

None

data Payload

The payload of the request

Constructors

Json Value

NonJson ByteString

Empty

data RequestException

The result when request fails.

Constructors

DecodeFailure ByteString String
JSON decoding the given response data failed.

ClientError Value
The HTTP response status code indicated failure.

RawClientError ByteString
The HTTP response status code indicated failure and the response was not valid JSON.

HttpException HttpExceptionContent
A wild exception upon sending the request

Lens

walletId :: HasType (ApiT WalletId) s => Lens' s Text

Constants

minUTxOValue :: ApiEra -> Natural

Minimum UTxO parameter for the test cluster.

The value returned by this function is only appropriate for a very minimal output, where:

the output has no assets other than ada.

the output uses a Shelley-era address of the length typically used in the integration test suite.

the output has no datum hash.

This value will almost certainly not be correct for outputs with non-ada assets, for outputs with longer addresses, or outputs with a datum hash.

In those cases, a larger value will be required. The precise value can be determined by calling one of the endpoints that returns an ApiFee object, and inspecting the minimumCoins field.

minUTxOValueForMinLengthAddress :: ApiEra -> Natural

slotLengthValue :: NominalDiffTime

Parameter in test cluster shelley genesis.

This space left blank intentionally.

securityParameterValue :: Word32

Parameter in test cluster shelley genesis.

epochLengthValue :: Word32

Parameter in test cluster shelley genesis.

defaultTxTTL :: NominalDiffTime

Wallet server's chosen transaction TTL value (in seconds) when none is given.

maximumCollateralInputCountByEra :: ApiEra -> Word16

minimumCollateralPercentageByEra :: ApiEra -> Natural

Create wallets

restoreWalletFromPubKey :: forall w (style :: WalletStyle) m. (Discriminate style, PostWallet style, HasType (ApiT WalletId) w, HasType (ApiT SyncProgress) w, Show w, FromJSON w, MonadIO m, MonadUnliftIO m) => Context -> Text -> Text -> ResourceT m w

Restore HW Wallet from pub key

emptyRandomWallet :: (MonadIO m, MonadUnliftIO m) => Context -> ResourceT m ApiByronWallet

Create an empty wallet

emptyRandomWalletMws :: (MonadIO m, MonadUnliftIO m) => Context -> ResourceT m (ApiByronWallet, Mnemonic 12)

emptyRandomWalletWithPasswd :: (MonadIO m, MonadUnliftIO m) => Context -> Text -> ResourceT m ApiByronWallet

emptyIcarusWallet :: (MonadIO m, MonadUnliftIO m) => Context -> ResourceT m ApiByronWallet

emptyIcarusWalletMws :: (MonadIO m, MonadUnliftIO m) => Context -> ResourceT m (ApiByronWallet, Mnemonic 15)

emptyByronWalletWith :: forall m. (MonadIO m, MonadUnliftIO m) => Context -> String -> (Text, [Text], Text) -> ResourceT m ApiByronWallet

postWallet :: (MonadIO m, MonadUnliftIO m) => Context -> Payload -> ResourceT m (Status, Either RequestException ApiWallet)

postWallet' :: (MonadIO m, MonadUnliftIO m) => Context -> Headers -> Payload -> ResourceT m (Status, Either RequestException ApiWallet)

postByronWallet :: (MonadIO m, MonadUnliftIO m) => Context -> Payload -> ResourceT m (Status, Either RequestException ApiByronWallet)

emptyWallet :: (MonadIO m, MonadUnliftIO m) => Context -> ResourceT m ApiWallet

Create an empty wallet

emptyWalletWith :: (MonadIO m, MonadUnliftIO m) => Context -> (Text, Text, Int) -> ResourceT m ApiWallet

Create an empty wallet

emptyWalletAndMnemonic :: MonadUnliftIO m => Context -> ResourceT m (ApiWallet, [Text])

emptyWalletAndMnemonicAndSndFactor :: MonadUnliftIO m => Context -> ResourceT m (ApiWallet, [Text], [Text])

emptyWalletAndMnemonicWith :: (MonadIO m, MonadUnliftIO m) => Context -> (Text, Text, Int) -> ResourceT m (ApiWallet, [Text])

emptyByronWalletFromXPrvWith :: forall m. (MonadIO m, MonadUnliftIO m) => Context -> String -> (Text, Text, Text) -> ResourceT m ApiByronWallet

rewardWallet :: (MonadIO m, MonadUnliftIO m) => Context -> ResourceT m (ApiWallet, Mnemonic 24)

postSharedWallet :: (MonadIO m, MonadUnliftIO m) => Context -> Headers -> Payload -> ResourceT m (Status, Either RequestException ApiSharedWallet)

deleteSharedWallet :: forall m. (MonadIO m, MonadUnliftIO m) => Context -> ApiSharedWallet -> m (Status, Either RequestException Value)

getSharedWallet :: forall m. (MonadIO m, MonadUnliftIO m) => Context -> ApiSharedWallet -> m (Status, Either RequestException ApiSharedWallet)

patchSharedWallet :: forall m. (MonadIO m, MonadUnliftIO m) => Context -> ApiSharedWallet -> CredentialType -> Payload -> m (Status, Either RequestException ApiSharedWallet)

getSharedWalletKey :: forall m. (MonadIO m, MonadUnliftIO m) => Context -> ApiSharedWallet -> Role -> DerivationIndex -> Maybe Bool -> m (Status, Either RequestException ApiVerificationKeyShared)

postAccountKeyShared :: forall m. (MonadIO m, MonadUnliftIO m) => Context -> ApiSharedWallet -> DerivationIndex -> Headers -> Payload -> m (Status, Either RequestException ApiAccountKeyShared)

getAccountKeyShared :: forall m. (MonadIO m, MonadUnliftIO m) => Context -> ApiSharedWallet -> Maybe KeyFormat -> m (Status, Either RequestException ApiAccountKeyShared)

getSomeVerificationKey :: forall m. (MonadIO m, MonadUnliftIO m) => Context -> ApiWallet -> m (ApiVerificationKeyShelley, ApiT (Hash "VerificationKey"))

Wallet helpers

listFilteredWallets

Arguments

:: (MonadIO m, MonadUnliftIO m)

=> Set Text
Set of walletIds to include

-> Context

-> m (Status, Either RequestException [ApiWallet])

Calls 'GET /wallets' and filters the response. This allows tests to be written for a parallel setting.

listFilteredByronWallets

Arguments

:: (MonadIO m, MonadUnliftIO m)

=> Set Text
Set of walletIds to include

-> Context

-> m (Status, Either RequestException [ApiByronWallet])

Calls 'GET /byron-wallets' and filters the response. This allows tests to be written for a parallel setting.

listFilteredSharedWallets

Arguments

:: (MonadIO m, MonadUnliftIO m)

=> Set Text
Set of walletIds to include

-> Context

-> m (Status, Either RequestException [ApiSharedWallet])

getWalletIdFromSharedWallet :: ApiSharedWallet -> ApiT WalletId

Helpers

(</>) :: ToHttpApiData a => Text -> a -> Text infixr 5

(!!) :: [a] -> Int -> a infixl 9

List index (subscript) operator, starting from 0. It is an instance of the more general genericIndex, which takes an index of any integral type.

computeApiCoinSelectionFee :: ApiCoinSelection n -> Coin

Computes the effective fee for an ApiCoinSelection value.

The ApiCoinSelection type doesn't include a field to indicate the fee.

This function computes the effective fee by subtracting the total output value from the total input value.

isValidDerivationPath :: Index 'Hardened 'PurposeK -> NonEmpty (ApiT DerivationIndex) -> Bool

isValidRandomDerivationPath :: NonEmpty (ApiT DerivationIndex) -> Bool

genMnemonics :: MnemonicLength -> IO [Text]

genMnemonics' :: forall mw ent csz m. (ConsistentEntropy ent mw csz, ValidEntropySize ent, ValidChecksumSize ent csz, ent ~ EntropySize mw, mw ~ MnemonicWords ent, MonadIO m) => m [Text]

getFromResponse :: HasCallStack => Lens' s a -> (Status, Either RequestException s) -> a

getFromResponseList :: Int -> Lens' s a -> (Status, Either RequestException [s]) -> a

json :: QuasiQuoter

joinStakePool :: forall n w m. (HasType (ApiT WalletId) w, DecodeAddress n, DecodeStakeAddress n, MonadIO m, MonadUnliftIO m) => Context -> ApiT PoolId -> (w, Text) -> m (Status, Either RequestException (ApiTransaction n))

joinStakePoolUnsigned :: forall n style w. (HasType (ApiT WalletId) w, DecodeAddress n, EncodeAddress n, DecodeStakeAddress n, Discriminate style) => Context -> w -> ApiT PoolId -> IO (Status, Either RequestException (ApiCoinSelection n))

delegationFee :: forall w m. (HasType (ApiT WalletId) w, MonadIO m, MonadUnliftIO m) => Context -> w -> m (Status, Either RequestException ApiFee)

quitStakePool :: forall n w m. (HasType (ApiT WalletId) w, DecodeAddress n, DecodeStakeAddress n, MonadIO m, MonadUnliftIO m) => Context -> (w, Text) -> m (Status, Either RequestException (ApiTransaction n))

quitStakePoolUnsigned :: forall n style w m. (HasType (ApiT WalletId) w, DecodeAddress n, DecodeStakeAddress n, EncodeAddress n, MonadIO m, Discriminate style) => Context -> w -> m (Status, Either RequestException (ApiCoinSelection n))

selectCoins :: forall n style w m. (HasType (ApiT WalletId) w, DecodeAddress n, DecodeStakeAddress n, EncodeAddress n, Discriminate style, MonadIO m, MonadUnliftIO m) => Context -> w -> NonEmpty (AddressAmount (ApiT Address, Proxy n)) -> m (Status, Either RequestException (ApiCoinSelection n))

selectCoinsWith :: forall n style w m. (HasType (ApiT WalletId) w, DecodeAddress n, DecodeStakeAddress n, EncodeAddress n, Discriminate style, MonadIO m, MonadUnliftIO m) => Context -> w -> NonEmpty (AddressAmount (ApiT Address, Proxy n)) -> (Payload -> Payload) -> m (Status, Either RequestException (ApiCoinSelection n))

listAddresses :: forall n m. (MonadIO m, MonadUnliftIO m, DecodeAddress n) => Context -> ApiWallet -> m [ApiAddress n]

signTx :: MonadUnliftIO m => Context -> ApiWallet -> ApiT SealedTx -> [(Status, Either RequestException ApiSerialisedTransaction) -> m ()] -> m ApiSerialisedTransaction

submitTx :: MonadUnliftIO m => Context -> ApiSerialisedTransaction -> [(Status, Either RequestException ApiTxId) -> m ()] -> m ApiTxId

submitTxWithWid :: MonadUnliftIO m => Context -> ApiWallet -> ApiSerialisedTransaction -> m (Status, Either RequestException ApiTxId)

getWallet :: forall w m. (MonadIO m, MonadUnliftIO m, HasType (ApiT WalletId) w) => Context -> w -> m ApiWallet

listTransactions :: forall n w m. (DecodeAddress n, DecodeStakeAddress n, HasType (ApiT WalletId) w, MonadIO m, MonadUnliftIO m) => Context -> w -> Maybe UTCTime -> Maybe UTCTime -> Maybe SortOrder -> m [ApiTransaction n]

listAllTransactions :: forall n w m. (DecodeAddress n, DecodeStakeAddress n, HasType (ApiT WalletId) w, MonadIO m, MonadUnliftIO m) => Context -> w -> m [ApiTransaction n]

deleteAllWallets :: Context -> IO ()

Delete all wallets

fixtureRawTx :: Context -> (Address, Natural) -> IO ByteString

fixtureRandomWallet :: (MonadIO m, MonadUnliftIO m) => Context -> ResourceT m ApiByronWallet

fixtureRandomWalletMws :: (MonadIO m, MonadUnliftIO m) => Context -> ResourceT m (ApiByronWallet, Mnemonic 12)

Restore a faucet Random wallet and wait until funds are available.

fixtureRandomWalletAddrs :: forall (n :: NetworkDiscriminant) m. (PaymentAddress n ByronKey, MonadIO m, MonadUnliftIO m) => Context -> ResourceT m (ApiByronWallet, [Address])

fixtureRandomWalletWith :: forall (n :: NetworkDiscriminant) m. (EncodeAddress n, DecodeAddress n, DecodeStakeAddress n, PaymentAddress n ByronKey, MonadIO m, MonadUnliftIO m) => Context -> [Natural] -> ResourceT m ApiByronWallet

Restore a wallet with the given UTxO distribution. Note that there's a limitation to what can be done here. We only have 10 UTxO available in each faucet and they "only" have 'faucetUtxoAmt = 100_000 Ada' in each.

This function makes no attempt at ensuring the request is valid, so be careful.

TODO: Remove duplication between Shelley / Byron fixtures.

fixtureIcarusWallet :: (MonadIO m, MonadUnliftIO m) => Context -> ResourceT m ApiByronWallet

fixtureIcarusWalletMws :: (MonadIO m, MonadUnliftIO m) => Context -> ResourceT m (ApiByronWallet, Mnemonic 15)

Restore a faucet Icarus wallet and wait until funds are available.

fixtureIcarusWalletAddrs :: forall (n :: NetworkDiscriminant) m. (PaymentAddress n IcarusKey, MonadIO m, MonadUnliftIO m) => Context -> ResourceT m (ApiByronWallet, [Address])

fixtureIcarusWalletWith :: forall (n :: NetworkDiscriminant) m. (EncodeAddress n, DecodeAddress n, DecodeStakeAddress n, PaymentAddress n IcarusKey, MonadIO m, MonadUnliftIO m) => Context -> [Natural] -> ResourceT m ApiByronWallet

Restore a wallet with the given UTxO distribution. Note that there's a limitation to what can be done here. We only have 10 UTxO available in each faucet and they "only" have 'faucetUtxoAmt = 100_000 Ada' in each.

This function makes no attempt at ensuring the request is valid, so be careful.

TODO: Remove duplication between Shelley / Byron fixtures.

fixtureWallet :: MonadIO m => Context -> ResourceT m ApiWallet

Restore a faucet and wait until funds are available.

Note: ResourceT is used to allow automatic garbage collection of unused wallets through small blocks of runResourceT (e.g. once per test). It doesn't return ReleaseKey since manual releasing is not needed.

fixtureWalletWith :: forall n m. (EncodeAddress n, DecodeAddress n, DecodeStakeAddress n, MonadIO m, MonadUnliftIO m) => Context -> [Natural] -> ResourceT m ApiWallet

Restore a wallet with the given UTxO distribution. Note that there's a limitation to what can be done here. We only have 10 UTxO available in each faucet and they "only" have 'faucetUtxoAmt = 100_000 Ada' in each.

This function makes no attempt at ensuring the request is valid, so be careful.

fixtureWalletWithMnemonics :: forall m scheme. (MonadIO m, NextWallet scheme) => Proxy scheme -> Context -> ResourceT m (ApiWallet, [Text])

fixtureMultiAssetWallet :: MonadIO m => Context -> ResourceT m ApiWallet

fixtureMultiAssetRandomWallet :: forall n m. (DecodeAddress n, DecodeStakeAddress n, EncodeAddress n, MonadIO m, MonadUnliftIO m) => Context -> ResourceT m ApiByronWallet

fixtureMultiAssetIcarusWallet :: forall n m. (DecodeAddress n, DecodeStakeAddress n, EncodeAddress n, MonadIO m, MonadUnliftIO m) => Context -> ResourceT m ApiByronWallet

constFixtureWalletNoWait :: MonadIO m => Context -> ResourceT m ApiWallet

Create a fixture from the same mnemonic every time. Don't wait for it to restore before returning.

faucetAmt :: Natural

Total amount on each faucet wallet

faucetUtxoAmt :: Natural

Each faucet wallet is composed of 10 times a single faucet UTxO of 100_000 Ada.

proc' :: FilePath -> [String] -> CreateProcess

postTx :: forall n w m. (DecodeAddress n, DecodeStakeAddress n, EncodeAddress n, HasType (ApiT WalletId) w, MonadIO m, MonadUnliftIO m) => Context -> (w, w -> (Method, Text), Text) -> ApiWallet -> Natural -> m (Status, Either RequestException (ApiTransaction n))

pickAnAsset :: TokenMap -> ((Text, Text), Natural)

mkTxPayloadMA :: forall n m. (DecodeAddress n, DecodeStakeAddress n, EncodeAddress n, MonadUnliftIO m) => (ApiT Address, Proxy n) -> Natural -> [((Text, Text), Natural)] -> Text -> m Payload

waitForServer :: forall ctx m. (HasType (Port "wallet") ctx, MonadIO m) => ctx -> m ()

Wait for a booting wallet server to start. Wait up to 30s or fail.

for :: [a] -> (a -> b) -> [b]

map flipped.

utcIso8601ToText :: UTCTime -> Text

eventually :: MonadIO m => String -> IO a -> m a

eventuallyUsingDelay

Arguments

:: MonadIO m

=> Int
Delay in microseconds

-> Int
Timeout in seconds

-> String
Brief description of the IO action

-> IO a

-> m a

fixturePassphrase :: Text

Default passphrase used for fixture wallets

fixturePassphraseEncrypted :: Text

fixturePassphrase encrypted by Scrypt function

waitForNextEpoch :: (MonadIO m, MonadUnliftIO m) => Context -> m ()

waitForTxImmutability :: MonadIO m => Context -> m ()

waitAllTxsInLedger :: forall n m. (DecodeAddress n, DecodeStakeAddress n, MonadIO m, MonadUnliftIO m) => Context -> ApiWallet -> m ()

toQueryString :: [(Text, Text)] -> Text

withMethod :: Method -> (Method, Text) -> (Method, Text)

Override the method of a particular endpoint, mostly to exercise invalid endpoints from existing ones.

withPathParam :: Int -> (Text -> Text) -> (Method, Text) -> (Method, Text)

Modifies the value of a path parameter at position n (indexed from 0) with the given update function. Throws if the given endpoint has no path parameter in the given position.

>>> Link.getWallet @Shelley myWallet
( "GET", "v2/wallets/2512a00e9653fe49a44a5886202e24d77eeb998f" )

>>> withPathParam 0 (<> "suffix") $ Link.getWallet @Shelley myWallet
( "GET", "v2/wallets/2512a00e9653fe49a44a5886202e24d77eeb998fsuffix" )

>>> withPathParam 1 (const "suffix") $ Link.joinStakePool @Shelley myPool myWallet
( "GET", "v2/stake-pools/2512a00e9653fe49a44a5886202e24d77eeb998f/wallets/patate" )

addField :: ToJSON a => Text -> a -> Payload -> Payload

icarusAddresses :: forall (n :: NetworkDiscriminant). PaymentAddress n IcarusKey => Mnemonic 15 -> [Address]

Generate an infinite list of addresses for icarus wallets

To be typically used as:

>>> take 1 (icarusAddresses @n)
[addr]

randomAddresses :: forall (n :: NetworkDiscriminant). PaymentAddress n ByronKey => Mnemonic 12 -> [Address]

Generate an infinite list of addresses for random wallets.

To be typically used as:

>>> take 1 (randomAddresses @n)
[addr]

shelleyAddresses :: forall (n :: NetworkDiscriminant). PaymentAddress n ShelleyKey => Mnemonic 15 -> [Address]

Generate an infinite list of addresses for shelley wallets.

To be typically used as:

>>> take 1 (shelleyAddresses @n)
[addr]

pubKeyFromMnemonics :: [Text] -> Text

rootPrvKeyFromMnemonics :: [Text] -> Text -> Text

unsafeGetTransactionTime :: MonadUnliftIO m => [ApiTransaction n] -> m UTCTime

getTxId :: ApiTransaction n -> String

oneSecond :: Int

getTTLSlots :: MonadIO m => Context -> NominalDiffTime -> m SlotNo

Converts a transaction TTL in seconds into a number of slots, using the slot length.

updateMetadataSource :: (MonadIO m, MonadUnliftIO m) => Context -> Text -> m ()

bracketSettings :: (MonadIO m, MonadUnliftIO m) => Context -> m () -> m ()

verifyMetadataSource :: (MonadIO m, MonadUnliftIO m) => Context -> PoolMetadataSource -> m ()

triggerMaintenanceAction :: (MonadIO m, MonadUnliftIO m) => Context -> Text -> m ()

verifyMaintenanceAction :: (MonadIO m, MonadUnliftIO m) => Context -> PoolMetadataGCStatus -> m ()

genXPubs :: Int -> IO [(XPub, Text)]

hexText :: ByteString -> Text

fromHexText :: Text -> Either String ByteString

accPubKeyFromMnemonics :: SomeMnemonic -> Maybe SomeMnemonic -> Word32 -> Passphrase "encryption" -> Text

Delegation helpers

notDelegating

Arguments

:: [(Maybe (ApiT PoolId), ApiEpochInfo)]
Pools to be joined & epoch at which the new delegation will become active

-> ApiWalletDelegation

Wallet not delegating and not about to join any stake pool.

delegating

Arguments

:: ApiT PoolId
Pool joined

-> [(Maybe (ApiT PoolId), ApiEpochInfo)]
Pools to be joined & epoch at which the new delegation will become active

-> ApiWalletDelegation

getSlotParams :: MonadIO m => Context -> m (EpochNo, SlottingParameters)

arbitraryStake :: Maybe Coin

CLI

commandName :: String

command :: CmdResult r => [CmdOption] -> String -> [String] -> IO r

Execute a system command. Before running command make sure you need any files that are required by the command.

This function takes a list of options (often just [], see CmdOption for the available options), the name of the executable (either a full name, or a program on the $PATH) and a list of arguments. The result is often (), but can be a tuple containg any of Stdout, Stderr and Exit. Some examples:

command_ [] "gcc" ["-c","myfile.c"]                          -- compile a file, throwing an exception on failure
Exit c <- command [] "gcc" ["-c",myfile]                     -- run a command, recording the exit code
(Exit c, Stderr err) <- command [] "gcc" ["-c","myfile.c"]   -- run a command, recording the exit code and error output
Stdout out <- command [] "gcc" ["-MM","myfile.c"]            -- run a command, recording the output
command_ [Cwd "generated"] "gcc" ["-c",myfile]               -- run a command in a directory

Unless you retrieve the ExitCode using Exit, any ExitFailure will throw an error, including the Stderr in the exception message. If you capture the Stdout or Stderr, that stream will not be echoed to the console, unless you use the option EchoStdout or EchoStderr.

If you use command inside a do block and do not use the result, you may get a compile-time error about being unable to deduce CmdResult. To avoid this error, use command_.

cardanoWalletCLI :: forall r m. (CmdResult r, MonadIO m) => [String] -> m r

Run a command using the 'cardano-wallet' executable.

generateMnemonicsViaCLI :: forall r m. (CmdResult r, MonadIO m) => [String] -> m r

createWalletViaCLI :: forall s m. (HasType (Port "wallet") s, MonadIO m) => s -> [String] -> String -> String -> String -> ResourceT m (ExitCode, String, Text)

createWalletFromPublicKeyViaCLI

Arguments

:: forall s m. (HasType (Port "wallet") s, MonadIO m)

=> s

-> [String]
NAME, [--address-pool-gap INT], ACCOUNT_PUBLIC_KEY

-> ResourceT m (Exit, Stdout, Stderr)

deleteWalletViaCLI :: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m) => s -> String -> m r

getWalletUtxoSnapshotViaCLI :: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m) => s -> String -> m r

getWalletUtxoStatisticsViaCLI :: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m) => s -> String -> m r

getWalletViaCLI :: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m) => s -> String -> m r

createAddressViaCLI

Arguments

:: forall s m. (HasType (Port "wallet") s, MonadIO m)

=> s

-> [String]
Args

-> String
Pass

-> m (ExitCode, Text, Text)
(ExitCode, StdOut, StdErr)

importAddressViaCLI

Arguments

:: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m)

=> s

-> [String]
Args

-> m r
(ExitCode, StdOut, StdErr)

listAddressesViaCLI :: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m) => s -> [String] -> m r

listStakePoolsViaCLI :: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m) => s -> m r

listWalletsViaCLI :: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m) => s -> m r

updateWalletNameViaCLI :: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m) => s -> [String] -> m r

updateWalletPassphraseViaCLI

Arguments

:: forall s m. (HasType (Port "wallet") s, MonadIO m)

=> s

-> String
Wallet id

-> String
Old passphrase

-> String
New passphrase

-> String
New passphrase (repeated for confirmation)

-> m (ExitCode, Text, Text)

updateWalletPassphraseWithMnemonicViaCLI

Arguments

:: forall s m. (HasType (Port "wallet") s, MonadIO m)

=> s

-> String
Wallet id

-> [Text]
Mnemonic

-> String
New passphrase

-> String
New passphrase (repeated for confirmation)

-> m (ExitCode, Text, Text)

postTransactionViaCLI :: forall s m. (HasType (Port "wallet") s, MonadIO m) => s -> String -> [String] -> m (ExitCode, String, Text)

postTransactionFeeViaCLI :: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m) => s -> [String] -> m r

listTransactionsViaCLI :: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m) => s -> TxMetadataSchema -> [String] -> m r

postExternalTransactionViaCLI :: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m) => s -> [String] -> m r

deleteTransactionViaCLI :: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m) => s -> String -> String -> m r

getTransactionViaCLI :: forall r s m. (CmdResult r, HasType (Port "wallet") s, MonadIO m) => s -> String -> String -> TxMetadataSchema -> m r

getRetirementEpoch :: ApiStakePool -> Maybe EpochNo

replaceStakeKey :: forall (n :: NetworkDiscriminant). (DecodeAddress n, EncodeAddress n) => (ApiT Address, Proxy n) -> (ApiT Address, Proxy n) -> (ApiT Address, Proxy n)

Returns the first address, modified to have the same stake key as the second address.

Only intended to be used with well-known inputs in tests, so throws if anything goes unexpectedly.

Re-exports

runResourceT :: MonadUnliftIO m => ResourceT m a -> m a

Unwrap a ResourceT transformer, and call all registered release actions.

Note that there is some reference counting involved due to resourceForkIO. If multiple threads are sharing the same collection of resources, only the last call to runResourceT will deallocate the resources.

NOTE Since version 1.2.0, this function will throw a ResourceCleanupException if any of the cleanup functions throw an exception.

Since: resourcet-0.3.0

data ResourceT (m :: Type -> Type) a

The Resource transformer. This transformer keeps track of all registered actions, and calls them upon exit (via runResourceT). Actions may be registered via register, or resources may be allocated atomically via allocate. allocate corresponds closely to bracket.

Releasing may be performed before exit via the release function. This is a highly recommended optimization, as it will ensure that scarce resources are freed early. Note that calling release will deregister the action, so that a release action will only ever be called once.

Since 0.3.0