Test.Integration.Framework.Request

Documentation

request

Arguments

:: forall a m s. (FromJSON a, MonadIO m, MonadUnliftIO m, HasType (URI, Manager) s)

=> s

-> (Method, Text)
HTTP method and request path

-> Headers
Request headers

-> Payload
Request body

-> m (Status, Either RequestException a)

Makes a request to the API and decodes the response.

rawRequest

Arguments

:: forall m s. (MonadIO m, MonadUnliftIO m, HasType (URI, Manager) s)

=> s

-> (Method, Text)
HTTP method and request path

-> Headers
Request headers

-> Payload
Request body

-> m (Status, Either RequestException ByteString)

Like request, but does not attempt to deserialize the response.

unsafeRequest :: forall a m. (FromJSON a, MonadIO m, MonadUnliftIO m) => Context -> (Method, Text) -> Payload -> m (Status, a)

Makes a request to the API, but throws if it fails.

data Headers

The headers of the request

Constructors

Headers RequestHeaders

Default

None

data Payload

The payload of the request

Constructors

Json Value

NonJson ByteString

Empty

data RequestException

The result when request fails.

Constructors

DecodeFailure ByteString String
JSON decoding the given response data failed.

ClientError Value
The HTTP response status code indicated failure.

RawClientError ByteString
The HTTP response status code indicated failure and the response was not valid JSON.

HttpException HttpExceptionContent
A wild exception upon sending the request