Test.Integration.Plutus

Documentation

mkSignerPolicy :: Value -> (Text, ApiT (Hash "TokenPolicy"))

Create a policy for which the only validation condition is that the transaction is signed by some key.

This template has one parameter:

vkHash: flat-encoded blake2b-224 hash of some verification key, in base16.

alwaysTrueValidator :: (Text, ApiT (Hash "TokenPolicy"))

pingPong_1 :: Value

pingPong_2 :: (MonadUnliftIO m, MonadFail m) => Value -> m Value

game_1 :: Value

game_2 :: (MonadUnliftIO m, MonadFail m) => Value -> m Value

game_3 :: (MonadUnliftIO m, MonadFail m) => Value -> m Value

mintBurn_1 :: (MonadUnliftIO m, MonadFail m) => Value -> m Value

A first transaction template which mints some token aimed for the wallet submitting the transaction (collected as part of balancing). Other than the minted token, the transaction has no inputs and no outputs. So the wallet is expected to balance it out and assign the minted token to a change address.

The template has three parameters:

policyId: A base16 policyId

policy: A base16 corresponding policy (see mkSignerPolicy)

vkHash: The verification key hash (base16) which was used to generate the policy.

mintBurn_2 :: (MonadUnliftIO m, MonadFail m) => Value -> m Value

withdrawScript_1 :: (MonadUnliftIO m, MonadFail m) => m Value

In this scenario, we attempt to perform a withdrawal of 42 ₳ from stake address locked by the special alwaysTrueValidator, which always validate.

This however requires to pass a Rewarding type of redeemer, which is quite exotic :)

currencyTx

Arguments

:: forall n. EncodeAddress n

=> ApiWalletInput n
UTxO that is hard-wired into the smart contract.

-> Value

Partial transaction that mints assets using the "currency" contract.

The contract is implemented in Plutus here:

https://github.com/input-output-hk/plutus-apps/blob/main/plutus-use-cases/src/Plutus/Contracts/Currency.hs

https://github.com/input-output-hk/plutus-apps/blob/main/plutus-use-cases/test/Spec/Currency.hs

The contract was serialised using the support code in extraPlutus/CurrencyContract.hs in this repository.

Assets can be minted by spending a specific UTxO that is hard-wired into the smart contract. Hence, minting can happen only once, as UTxOs can be spent only once.

The transaction is partial and needs to be sent to the /transactions-balance endpoint of the wallet in order to add change outputs and further inputs to cover fees if necessary. This endpoint also needs to make sure that the minted assets are assigned to a change output.

toHex :: ByteString -> Text