Test.Integration.Faucet

Documentation

newtype Faucet

An opaque Faucet type from which one can get a wallet with funds

Constructors

Faucet (MVar [Mnemonic 15])

nextWallet :: Faucet -> IO (Mnemonic 15)

Get the next faucet wallet. Requires the initFaucet to be called in order to get a hand on a Faucet.