Test.Integration.Framework.DSL

Documentation

data Context t

Running Context for our integration test

Constructors

Context

Fields

_cleanup :: IO ()

An action to clean up open processes after tests

_manager :: (Text, Manager)

The underlying BaseUrl and Manager used by the Wallet Client

_walletPort :: Port "wallet"

Server TCP port

_nodePort :: Port "node"

Jormungandr REST API port

_faucet :: Faucet

A Faucet handle in to have access to funded wallets in integration tests.

_feeEstimator :: TxDescription -> (Natural, Natural)

A fee estimator for the integration tests

_target :: Proxy t

class KnownCommand t where

A class to select the right command for a given 'Context t'

Methods

commandName :: String

data TxDescription

Describe a transaction in terms of its inputs and outputs

Constructors

TxDescription

Fields

nInputs :: Int

nOutputs :: Int

Steps

request

Arguments

:: (FromJSON a, MonadIO m, MonadCatch m, HasType (Text, Manager) s)

=> s

-> (Method, Text)
HTTP method and request path

-> Headers
Request headers

-> Payload
Request body

-> m (Status, Either RequestException a)

Makes a request to the API and decodes the response.

unsafeRequest :: forall a m t. (FromJSON a, MonadIO m, MonadCatch m) => Context t -> (Method, Text) -> Payload -> m (Status, a)

Makes a request to the API, but throws if it fails.

Expectations

expectPathEventuallyExist :: FilePath -> IO ()

A file is eventually created on the given location

expectSuccess :: (MonadIO m, MonadFail m) => (s, Either RequestException a) -> m ()

Expect a successful response, without any further assumptions

expectError :: (MonadIO m, MonadFail m, Show a) => (s, Either RequestException a) -> m ()

Expect an errored response, without any further assumptions

expectErrorMessage :: (MonadIO m, MonadFail m, Show a) => String -> (s, Either RequestException a) -> m ()

Expect an errored response, without any further assumptions

expectFieldEqual :: (MonadIO m, MonadFail m, Show a, Eq a) => Lens' s a -> a -> (Status, Either RequestException s) -> m ()

expectFieldNotEqual :: (MonadIO m, MonadFail m, Show a, Eq a) => Lens' s a -> a -> (Status, Either RequestException s) -> m ()

expectFieldBetween :: (MonadIO m, MonadFail m, Show a, Ord a) => Lens' s a -> (a, a) -> (Status, Either RequestException s) -> m ()

expectListItemFieldBetween :: (MonadIO m, MonadFail m, Show a, Eq a, Ord a) => Int -> Lens' s a -> (a, a) -> (Status, Either RequestException [s]) -> m ()

expectListItemFieldEqual :: (MonadIO m, MonadFail m, Show a, Eq a) => Int -> Lens' s a -> a -> (Status, Either RequestException [s]) -> m ()

Expects that returned data list's particular item field matches the expected value e.g. expectListItemFieldEqual 0 walletName "first" response expectListItemFieldEqual 1 walletName "second" response

expectListSizeEqual :: (MonadIO m, MonadFail m, Foldable xs) => Int -> (Status, Either RequestException (xs a)) -> m ()

Expects data list returned by the API to be of certain length

expectResponseCode :: MonadIO m => Status -> (Status, a) -> m ()

Expect a given response code on the response

expectEventually :: (MonadIO m, MonadCatch m, MonadFail m) => (Ord a, Show a) => HasType (Text, Manager) ctx => ctx -> Lens' ApiWallet a -> a -> (Status, Either RequestException ApiWallet) -> m ()

Expects wallet from the request to eventually reach the given state or beyond

expectEventually' :: (MonadIO m, MonadCatch m, MonadFail m) => (Ord a, Show a) => HasType (Text, Manager) ctx => ctx -> Lens' ApiWallet a -> a -> ApiWallet -> m ()

Same as expectEventually but work directly on ApiWallet , not response from the API

expectValidJSON :: forall m a. (MonadFail m, FromJSON a) => Proxy a -> String -> m a

Expects a given string to be a valid JSON output corresponding to some given data-type a. Returns this type if successful.

expectCliFieldBetween :: (MonadIO m, MonadFail m, Show a, Ord a) => Lens' s a -> (a, a) -> s -> m ()

expectCliFieldEqual :: (MonadIO m, Show a, Eq a) => Lens' s a -> a -> s -> m ()

expectCliFieldNotEqual :: (MonadIO m, Show a, Eq a) => Lens' s a -> a -> s -> m ()

expectCliListItemFieldBetween :: (MonadIO m, MonadFail m, Show a, Eq a, Ord a) => Int -> Lens' s a -> (a, a) -> [s] -> m ()

expectCliListItemFieldEqual :: (MonadIO m, MonadFail m, Show a, Eq a) => Int -> Lens' s a -> a -> [s] -> m ()

Same as expectListItemFieldEqual but for CLI

expectWalletUTxO :: (MonadIO m, MonadFail m) => [Word64] -> Either RequestException ApiUtxoStatistics -> m ()

Expects wallet UTxO statistics from the request to be equal to pre-calculated

verify :: Monad m => a -> [a -> m ()] -> m ()

Apply a to all actions in sequence

data Headers

The headers of the request

Constructors

Headers RequestHeaders

Default

None

data Payload

The payload of the request

Constructors

Json Value

NonJson ByteString

Empty

data RequestException

The result when request fails.

Constructors

DecodeFailure ByteString
JSON decoding the given response data failed.

ClientError Value
The HTTP response status code indicated failure.

HttpException HttpExceptionContent
A wild exception upon sending the request

Lens

addressPoolGap :: HasType (ApiT AddressPoolGap) s => Lens' s Int

amount :: HasType (Quantity "lovelace" Natural) s => Lens' s Natural

balanceAvailable :: HasType (ApiT WalletBalance) s => Lens' s Natural

balanceTotal :: HasType (ApiT WalletBalance) s => Lens' s Natural

delegation :: HasType (ApiT (WalletDelegation (ApiT PoolId))) s => Lens' s (WalletDelegation (ApiT PoolId))

direction :: HasType (ApiT Direction) s => Lens' s Direction

feeEstimator :: Lens' (Context t) (TxDescription -> (Natural, Natural))

inputs :: HasType [ApiTxInput t] s => Lens' s [ApiTxInput t]

passphraseLastUpdate :: HasType (Maybe (ApiT WalletPassphraseInfo)) s => Lens' s (Maybe Text)

state :: HasField' "state" s (ApiT t) => Lens' s t

status :: HasType (ApiT TxStatus) s => Lens' s TxStatus

outputs :: HasType (NonEmpty (AddressAmount t)) s => Lens' s [AddressAmount t]

walletId :: HasType (ApiT WalletId) s => Lens' s Text

walletName :: HasType (ApiT WalletName) s => Lens' s Text

Helpers

(</>) :: ToHttpApiData a => Text -> a -> Text infixr 5

(!!) :: [a] -> Int -> a infixl 9

List index (subscript) operator, starting from 0. It is an instance of the more general genericIndex, which takes an index of any integral type.

emptyWallet :: Context t -> IO ApiWallet

Create an empty wallet

emptyWalletWith :: Context t -> (Text, Text, Int) -> IO ApiWallet

Create an empty wallet

getFromResponse :: Lens' s a -> (Status, Either RequestException s) -> a

getFromResponseList :: Int -> Lens' s a -> (Status, Either RequestException [s]) -> a

getJormungandrBlock0H :: IO String

json :: QuasiQuoter

listAddresses :: forall t. DecodeAddress t => Context t -> ApiWallet -> IO [ApiAddress t]

listTransactions :: forall t. DecodeAddress t => Context t -> ApiWallet -> Maybe UTCTime -> Maybe UTCTime -> Maybe SortOrder -> IO [ApiTransaction t]

listAllTransactions :: forall t. DecodeAddress t => Context t -> ApiWallet -> IO [ApiTransaction t]

tearDown :: Context t -> IO ()

teardown after each test (currently only deleting all wallets)

fixtureWallet :: Context t -> IO ApiWallet

Restore a faucet and wait until funds are available.

fixtureWalletWith :: forall t. (EncodeAddress t, DecodeAddress t) => Context t -> [Natural] -> IO ApiWallet

Restore a wallet with the given UTxO distribution. Note that there's a limitation to what can be done here. We only have 10 UTxO available in each faucet and they "only" have 'faucetUtxoAmt = 100_000 Ada' in each.

This function makes no attempt at ensuring the request is valid, so be careful.

faucetAmt :: Natural

Total amount on each faucet wallet

faucetUtxoAmt :: Natural

Each faucet wallet is composed of 10 times a single faucet UTxO of 100_000 Ada.

proc' :: FilePath -> [String] -> CreateProcess

waitForServer :: forall t ctx. (HasType (Port "wallet") ctx, KnownCommand t) => ctx -> IO ()

Wait for a booting wallet server to start. Wait up to 30s or fail.

collectStreams :: (Int, Int) -> CreateProcess -> IO (Text, Text)

Collect lines from standard output and error streams for 30 seconds, or, until a given limit is for both streams.

shouldContainT :: Text -> Text -> IO ()

Like shouldContain, but with Text

shouldNotContainT :: Text -> Text -> IO ()

Like shouldNotContain, but with Text

for :: [a] -> (a -> b) -> [b]

map flipped.

toQueryString :: [(Text, Text)] -> Text

utcIso8601ToText :: UTCTime -> Text

prepExternalTxViaJcli :: Port "node" -> Text -> Natural -> IO Text

Prepare externally signed Tx for Jormungandr

Endpoints

getWalletEp :: ApiWallet -> (Method, Text)

deleteWalletEp :: ApiWallet -> (Method, Text)

getWalletUtxoEp :: ApiWallet -> (Method, Text)

getAddressesEp :: ApiWallet -> Text -> (Method, Text)

postTxEp :: ApiWallet -> (Method, Text)

postExternalTxEp :: (Method, Text)

postTxFeeEp :: ApiWallet -> (Method, Text)

listTxEp :: ApiWallet -> Text -> (Method, Text)

updateWalletPassEp :: ApiWallet -> (Method, Text)

CLI

runJcli :: CmdResult r => [String] -> IO r

command :: CmdResult r => [CmdOption] -> String -> [String] -> IO r

cardanoWalletCLI :: forall t r. (CmdResult r, KnownCommand t) => [String] -> IO r

Run a command using the 'cardano-wallet' executable. We run it through stack as we intend to also get code-coverage from running these commands!

generateMnemonicsViaCLI :: forall t r. (CmdResult r, KnownCommand t) => [String] -> IO r

createWalletViaCLI :: forall t s. (HasType (Port "wallet") s, KnownCommand t) => s -> [String] -> String -> String -> String -> IO (ExitCode, String, Text)

deleteWalletViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> String -> IO r

getWalletUtxoStatisticsViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> String -> IO r

getWalletViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> String -> IO r

listAddressesViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> [String] -> IO r

listWalletsViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> IO r

updateWalletNameViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> [String] -> IO r

updateWalletPassphraseViaCLI

Arguments

:: (KnownCommand t, HasType (Port "wallet") s)

=> s

-> String
Wallet id

-> String
Old passphrase

-> String
New passphrase

-> String
New passphrase (repeated for confirmation)

-> IO (ExitCode, Text, Text)

postTransactionViaCLI :: forall t s. (HasType (Port "wallet") s, KnownCommand t) => s -> String -> [String] -> IO (ExitCode, String, Text)

postTransactionFeeViaCLI :: forall t s. (HasType (Port "wallet") s, KnownCommand t) => s -> [String] -> IO (ExitCode, String, Text)

listTransactionsViaCLI :: forall t r s. (CmdResult r, HasType (Port "wallet") s, KnownCommand t) => s -> [String] -> IO r

postExternalTransactionViaCLI :: forall t r s. (CmdResult r, HasType (Port "wallet") s, KnownCommand t) => s -> [String] -> IO r