Test.Integration.Framework.DSL

Documentation

data Context t

Running Context for our integration test

Constructors

Context

Fields

_cleanup :: IO ()

Cleanup action for the context

_manager :: (Text, Manager)

The underlying BaseUrl and Manager used by the Wallet Client

_walletPort :: Port "wallet"

Server TCP port

_faucet :: Faucet

A Faucet handle in to have access to funded wallets in integration tests.

_feeEstimator :: TxDescription -> (Natural, Natural)

A fee estimator for the integration tests

_target :: Proxy t

class KnownCommand t where

A class to select the right command for a given 'Context t'

Methods

commandName :: String

data TxDescription

Describe a transaction in terms of its inputs and outputs

Constructors

DelegDescription

Fields

nInputs :: Int

nOutputs :: Int

nCertificates :: Int

PaymentDescription

Fields

nInputs :: Int

nOutputs :: Int

nChanges :: Int

Steps

request

Arguments

:: (FromJSON a, MonadIO m, MonadCatch m, HasType (Text, Manager) s)

=> s

-> (Method, Text)
HTTP method and request path

-> Headers
Request headers

-> Payload
Request body

-> m (Status, Either RequestException a)

Makes a request to the API and decodes the response.

unsafeRequest :: forall a m t. (FromJSON a, MonadIO m, MonadCatch m) => Context t -> (Method, Text) -> Payload -> m (Status, a)

Makes a request to the API, but throws if it fails.

Expectations

expectPathEventuallyExist :: HasCallStack => FilePath -> IO ()

A file is eventually created on the given location

expectSuccess :: (HasCallStack, MonadIO m, MonadFail m) => (s, Either RequestException a) -> m ()

Expect a successful response, without any further assumptions.

expectError :: (HasCallStack, MonadIO m, MonadFail m, Show a) => (s, Either RequestException a) -> m ()

Expect an error response, without any further assumptions.

expectErrorMessage :: (HasCallStack, MonadIO m, MonadFail m, Show a) => String -> (s, Either RequestException a) -> m ()

Expect an error response, without any further assumptions.

expectField :: (HasCallStack, MonadIO m, MonadFail m, Show a) => Lens' s a -> (a -> Expectation) -> (Status, Either RequestException s) -> m ()

expectListField :: (HasCallStack, MonadIO m, MonadFail m, Show a) => Int -> Lens' s a -> (a -> Expectation) -> (Status, Either RequestException [s]) -> m ()

expectListSize :: (HasCallStack, MonadIO m, MonadFail m, Foldable xs) => Int -> (Status, Either RequestException (xs a)) -> m ()

Expects data list returned by the API to be of certain length

expectResponseCode :: (HasCallStack, MonadIO m, Show a) => Status -> (Status, a) -> m ()

Expect a given response code on the response.

expectValidJSON :: forall m a. (HasCallStack, MonadFail m, FromJSON a) => Proxy a -> String -> m a

Expects a given string to be a valid JSON output corresponding to some given data-type a. Returns this type if successful.

expectCliField :: (HasCallStack, MonadIO m, MonadFail m, Show a) => Lens' s a -> (a -> Expectation) -> s -> m ()

expectCliListField :: (HasCallStack, MonadIO m, MonadFail m, Show a) => Int -> Lens' s a -> (a -> Expectation) -> [s] -> m ()

expectWalletUTxO :: (HasCallStack, MonadIO m, MonadFail m) => [Word64] -> Either RequestException ApiUtxoStatistics -> m ()

Expects wallet UTxO statistics from the request to be equal to pre-calculated statistics.

between :: (Ord a, Show a) => (a, a) -> a -> Expectation

(.>=) :: (Ord a, Show a) => a -> a -> Expectation

(.>) :: (Ord a, Show a) => a -> a -> Expectation

verify :: Monad m => a -> [a -> m ()] -> m ()

Apply a to all actions in sequence

data Headers

The headers of the request

Constructors

Headers RequestHeaders

Default

None

data Payload

The payload of the request

Constructors

Json Value

NonJson ByteString

Empty

data RequestException

The result when request fails.

Constructors

DecodeFailure ByteString
JSON decoding the given response data failed.

ClientError Value
The HTTP response status code indicated failure.

HttpException HttpExceptionContent
A wild exception upon sending the request

expectedBlockchainParams :: ApiNetworkParameters

Lens

walletId :: HasType (ApiT WalletId) s => Lens' s Text

Helpers

(</>) :: ToHttpApiData a => Text -> a -> Text infixr 5

(!!) :: [a] -> Int -> a infixl 9

List index (subscript) operator, starting from 0. It is an instance of the more general genericIndex, which takes an index of any integral type.

emptyRandomWallet :: Context t -> IO ApiByronWallet

Create an empty wallet

emptyIcarusWallet :: Context t -> IO ApiByronWallet

emptyByronWalletWith :: forall t. Context t -> String -> (Text, [Text], Text) -> IO ApiByronWallet

emptyWallet :: Context t -> IO ApiWallet

Create an empty wallet

emptyWalletWith :: Context t -> (Text, Text, Int) -> IO ApiWallet

Create an empty wallet

getFromResponse :: Lens' s a -> (Status, Either RequestException s) -> a

getFromResponseList :: Int -> Lens' s a -> (Status, Either RequestException [s]) -> a

json :: QuasiQuoter

joinStakePool :: forall t w. HasType (ApiT WalletId) w => Context t -> ApiT PoolId -> (w, Text) -> IO (Status, Either RequestException (ApiTransaction Testnet))

delegationFee :: forall t w. HasType (ApiT WalletId) w => Context t -> w -> IO (Status, Either RequestException ApiFee)

quitStakePool :: forall t w. HasType (ApiT WalletId) w => Context t -> (w, Text) -> IO (Status, Either RequestException (ApiTransaction Testnet))

selectCoins :: forall t w. HasType (ApiT WalletId) w => Context t -> w -> NonEmpty (AddressAmount Testnet) -> IO (Status, Either RequestException (ApiCoinSelection Testnet))

listAddresses :: Context t -> ApiWallet -> IO [ApiAddress Testnet]

listTransactions :: Context t -> ApiWallet -> Maybe UTCTime -> Maybe UTCTime -> Maybe SortOrder -> IO [ApiTransaction Testnet]

listAllTransactions :: Context t -> ApiWallet -> IO [ApiTransaction Testnet]

tearDown :: Context t -> IO ()

teardown after each test (currently only deleting all wallets)

fixtureRawTx :: Context t -> (Address, Natural) -> IO ByteString

fixtureRandomWallet :: Context t -> IO ApiByronWallet

Restore a faucet Random wallet and wait until funds are available.

fixtureIcarusWallet :: Context t -> IO ApiByronWallet

Restore a faucet Icarus wallet and wait until funds are available.

fixtureWallet :: Context t -> IO ApiWallet

Restore a faucet and wait until funds are available.

fixtureWalletWith :: Context t -> [Natural] -> IO ApiWallet

Restore a wallet with the given UTxO distribution. Note that there's a limitation to what can be done here. We only have 10 UTxO available in each faucet and they "only" have 'faucetUtxoAmt = 100_000 Ada' in each.

This function makes no attempt at ensuring the request is valid, so be careful.

fixtureWalletWithMnemonics :: Context t -> IO (ApiWallet, [Text])

faucetAmt :: Natural

Total amount on each faucet wallet

faucetUtxoAmt :: Natural

Each faucet wallet is composed of 10 times a single faucet UTxO of 100_000 Ada.

proc' :: FilePath -> [String] -> CreateProcess

waitForServer :: forall t ctx. (HasType (Port "wallet") ctx, KnownCommand t) => ctx -> IO ()

Wait for a booting wallet server to start. Wait up to 30s or fail.

for :: [a] -> (a -> b) -> [b]

map flipped.

utcIso8601ToText :: UTCTime -> Text

eventually :: String -> IO a -> IO a

eventuallyUsingDelay

Arguments

:: Int
Delay in microseconds

-> String
Brief description of the IO action

-> IO a

-> IO a

fixturePassphrase :: Text

Default passphrase used for fixture wallets

waitForNextEpoch :: Context t -> IO ()

waitAllTxsInLedger :: forall t n. n ~ Testnet => Context t -> ApiWallet -> IO ()

toQueryString :: [(Text, Text)] -> Text

withMethod :: Method -> (Method, Text) -> (Method, Text)

Override the method of a particular endpoint, mostly to exercise invalid endpoints from existing ones.

withPathParam :: Int -> (Text -> Text) -> (Method, Text) -> (Method, Text)

Modifies the value of a path parameter at position n (indexed from 0) with the given update function. Throws if the given endpoint has no path parameter in the given position.

>>> Link.getWallet @Shelley myWallet
( "GET", "v2/wallets/2512a00e9653fe49a44a5886202e24d77eeb998f" )

>>> withPathParam 0 (<> "suffix") $ Link.getWallet @Shelley myWallet
( "GET", "v2/wallets/2512a00e9653fe49a44a5886202e24d77eeb998fsuffix" )

>>> withPathParam 1 (const "suffix") $ Link.joinStakePool @Shelley myPool myWallet
( "GET", "v2/stake-pools/2512a00e9653fe49a44a5886202e24d77eeb998f/wallets/patate" )

Delegation helpers

mkEpochInfo

Arguments

:: EpochNo
Epoch to construct

-> SlotParameters
Blockchain slot parameters

-> ApiEpochInfo

Handy constructor for ApiEpochInfo

notDelegating

Arguments

:: [(Maybe (ApiT PoolId), ApiEpochInfo)]
Pools to be joined & epoch at which the new delegation will become active

-> ApiWalletDelegation

Wallet not delegating and not about to join any stake pool.

delegating

Arguments

:: ApiT PoolId
Pool joined

-> [(Maybe (ApiT PoolId), ApiEpochInfo)]
Pools to be joined & epoch at which the new delegation will become active

-> ApiWalletDelegation

getSlotParams :: Context t -> IO (EpochNo, SlotParameters)

CLI

command :: CmdResult r => [CmdOption] -> String -> [String] -> IO r

cardanoWalletCLI :: forall t r. (CmdResult r, KnownCommand t) => [String] -> IO r

Run a command using the 'cardano-wallet' executable for the target t.

generateMnemonicsViaCLI :: forall t r. (CmdResult r, KnownCommand t) => [String] -> IO r

createWalletViaCLI :: forall t s. (HasType (Port "wallet") s, KnownCommand t) => s -> [String] -> String -> String -> String -> IO (ExitCode, String, Text)

deleteWalletViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> String -> IO r

getWalletUtxoStatisticsViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> String -> IO r

getWalletViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> String -> IO r

listAddressesViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> [String] -> IO r

listStakePoolsViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> IO r

listWalletsViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> IO r

updateWalletNameViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> [String] -> IO r

updateWalletPassphraseViaCLI

Arguments

:: (KnownCommand t, HasType (Port "wallet") s)

=> s

-> String
Wallet id

-> String
Old passphrase

-> String
New passphrase

-> String
New passphrase (repeated for confirmation)

-> IO (ExitCode, Text, Text)

postTransactionViaCLI :: forall t s. (HasType (Port "wallet") s, KnownCommand t) => s -> String -> [String] -> IO (ExitCode, String, Text)

postTransactionFeeViaCLI :: forall t s. (HasType (Port "wallet") s, KnownCommand t) => s -> [String] -> IO (ExitCode, String, Text)

listTransactionsViaCLI :: forall t r s. (CmdResult r, HasType (Port "wallet") s, KnownCommand t) => s -> [String] -> IO r

postExternalTransactionViaCLI :: forall t r s. (CmdResult r, HasType (Port "wallet") s, KnownCommand t) => s -> [String] -> IO r

deleteTransactionViaCLI :: forall t r s. (CmdResult r, KnownCommand t, HasType (Port "wallet") s) => s -> String -> String -> IO r