Test.Integration.Framework.Request

Documentation

request

Arguments

:: (FromJSON a, MonadIO m, MonadCatch m, HasType (Text, Manager) s)

=> s

-> (Method, Text)
HTTP method and request path

-> Headers
Request headers

-> Payload
Request body

-> m (Status, Either RequestException a)

Makes a request to the API and decodes the response.

unsafeRequest :: forall a m t. (FromJSON a, MonadIO m, MonadCatch m) => Context t -> (Method, Text) -> Payload -> m (Status, a)

Makes a request to the API, but throws if it fails.

data Headers

The headers of the request

Constructors

Headers RequestHeaders

Default

None

data Payload

The payload of the request

Constructors

Json Value

NonJson ByteString

Empty

data RequestException

The result when request fails.

Constructors

DecodeFailure ByteString
JSON decoding the given response data failed.

ClientError Value
The HTTP response status code indicated failure.

HttpException HttpExceptionContent
A wild exception upon sending the request

data Context t

Running Context for our integration test

Constructors

Context

Fields

_cleanup :: IO ()

An action to clean up open processes after tests

_manager :: (Text, Manager)

The underlying BaseUrl and Manager used by the Wallet Client

_walletPort :: Port "wallet"

Server TCP port

_nodePort :: Port "node"

Jormungandr REST API port

_faucet :: Faucet

A Faucet handle in to have access to funded wallets in integration tests.

_feeEstimator :: TxDescription -> (Natural, Natural)

A fee estimator for the integration tests

_target :: Proxy t

data TxDescription

Describe a transaction in terms of its inputs and outputs

Constructors

TxDescription

Fields

nInputs :: Int

nOutputs :: Int